{"types":"C","runs":30}

+++ Problem21.java
@@ -7,7 +7,7 @@ public class Problem21 extends GraphicsProgram {
 		drawVerticalLines();
 	}
 	private void drawVerticalLines() {
-		double n = (double)getWidth() / 10;
+		double n = getWidth() / 10;
 		for (int x = 0; x <= getWidth(); x += n) {
 			GLine vLine = new GLine(x, 0, x, getHeight());
 			add(vLine);
@@ -16,7 +16,7 @@ public class Problem21 extends GraphicsProgram {
 		
 	}
 	private void drawHorizontalLines() {
-		double n = (double)getHeight() / 10;
+		double n = getHeight() / 10;
 		for (int y = 0; y <= getHeight(); y += n) {
 			GLine hLine = new GLine(0, y, getWidth(), y);
 			add(hLine);
